+++ cmd/.snapshots/TestCreateProjectFiles-dumb-main.py
@@ -1,29 +1,40 @@
 import cog_settings
 from data_pb2 import DumbAction
 
-from cogment import Agent, GrpcServer
+import cogment
 
-class Dumb(Agent):
-    VERSIONS = {"dumb": "1.0.0"}
+import asyncio
+
+async def dumb_impl(actor_session):
     actor_class = cog_settings.actor_classes.dumb
 
-    def decide(self, observation: cog_settings.actor_classes.dumb.observation_space):
-        print("Dumb decide")
-        action = DumbAction()
-        return action
+    actor_session.start()
 
-    def reward(self, reward):
-        print("Dumb reward")
+    async for event in actor_session.event_loop():
+        if "observation" in event:
+            observation = event["observation"]
+            print("Dumb decide")
+            action = DumbAction()
+            actor_session.do_action(action)
+        if "reward" in event:
+            reward = event["reward"]
+            print("Dumb reward")
+        if "message" in event:
+            (msg, sender) = event["message"]
+            print(f"Dumb received message - {msg} from sender {sender}")
 
-    def on_message(self, sender, msg):
-        if msg:
-            print(f'Agent {self.id_in_class} received message - {msg} from sender {sender}')
+async def main():
+    print("Dumb service up and running.")
 
-    def end(self):
-        print("Dumb end")
+    context = cogment.Context(cog_settings=cog_settings, user_id="testit")
 
+    context.register_actor(
+        impl=dumb_impl,
+        impl_name="dumb_impl",
+        actor_classes=["dumb"])
+
+    await context.serve_all_registered(port=9000)
 
 if __name__ == '__main__':
-    server = GrpcServer(Dumb, cog_settings)
-    server.serve()
+    asyncio.run(main())